Add record/data creation/view pages and routes
Pretty much implemented the core logic for v1.
Added views for listing all records and then for each record, listing data and allowing to add
then I built the routes for them, and the POST route allowing data to be added to a record.
Also removed the 'admin' account so that I can push to heroku